Feat: Added Launguage switcher & Notifications Badge to the context bar
