+++ resources/css/app.css
@@ -127,7 +127,7 @@
     --sidebar-accent-foreground: oklch(0.205 0 0);
     --sidebar-border: oklch(0.922 0 0);
     --sidebar-ring: oklch(0.87 0 0);
-    --fmpo: #08c9b7;
+    --fmpo: #098f81;
 }
 
 .dark {
