+++ resources/js/components/ControlCenterCard.jsx
@@ -12,12 +12,14 @@ export default function ControlCenterCard({ item }) {
     const description = translations?.[item.descriptionKey] || item.fallbackDescription;
 
     return (
-        // The Link wraps the Card, making the whole thing clickable
-        <Link href={route(item.route)} className="block outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background rounded-lg">
-            <Card className="h-full transition-all duration-200 hover:border-primary hover:shadow-lg hover:-translate-y-1">
+        <Link
+            href={route(item.route)}
+            className="focus-visible:ring-offset-background block rounded-lg outline-none focus-visible:ring-2 focus-visible:ring-[var(--fmpo)] focus-visible:ring-offset-2"
+        >
+            <Card className="h-full transition-all duration-100 hover:-translate-y-1 hover:border-[var(--fmpo)] hover:shadow-lg">
                 <CardHeader className="flex flex-row items-center gap-4 p-4">
-                    <div className="flex-shrink-0 p-3 ">
-                        <Icon icon={item.icon} className="w-8 h-8 text-primary" />
+                    <div className="flex-shrink-0 p-3">
+                        <Icon icon={item.icon} className="h-8 w-8 text-[var(--fmpo)]" />
                     </div>
                     <div className="flex-grow">
                         <CardTitle className="text-base">{title}</CardTitle>
